fix(alerts/tests): test_spec015_alert_retry_backoff.py float/int 불일치 2건
1. AlertEngine.__init__의 window_seconds: int = 60는 내부에서
   `now - self.window_seconds`(now는 time.time(), float) 형태로만
   쓰여 float도 완전히 안전한데, 이 테스트는 윈도우 만료를 빠르게
   검증하려고 window_seconds=0.05(50ms)를 실제로 필요로 한다
   (test_window_expiry_allows_new_alerts). float로 넓힘(engine.py
   소스 수정) — 다른 정수 인자와 무관, 순수 산술 컨텍스트라 동작 무변화.

2. 테스트 헬퍼 _make_rule(cooldown: float = 0.0)이 AlertRule.cooldown:
   int = 300로 forward하는데, 실제 호출부는 전부 cooldown=0.0 또는
   cooldown=300만 쓰고 진짜 소수 값이 필요한 곳은 없었다 — 헬퍼 타입을
   int = 0으로 맞추고 cooldown=0.0 → cooldown=0으로 정리(동작 무변화,
   실제 쓰임새에 맞는 정확한 타입).

pyright 0 errors 확인(engine.py + 테스트 파일 모두). 이 파일 13 passed,
alert 관련 전체 106 passed, 전체 스위트 3601 passed.

+++ agent_evaluator/alerts/engine.py
@@ -195,7 +195,7 @@ class AlertEngine:
         history_dir: str | None = None,
         async_dispatch: bool = False,
         max_alerts_per_window: int | None = None,
-        window_seconds: int = 60,
+        window_seconds: float = 60,
     ) -> None:
         self.history = AlertHistory(history_dir)
         self._rules: list[AlertRule] = []
